Enable warnings about integer conversions and fix them
Tasking compiler have some of those warnings enabled by default so let's
them enabled even for gcc.

It seems that fixing those warnings changed the functionality and I don't
have time to solve these problems right now.

+++ macan/src/common.c
@@ -25,6 +25,11 @@
 #include <stdint.h>
 #include <macan_private.h>
 
+/* TODO: A lot of functions (e.g. canid2ecuid) return either error as
+ * -1 or valid answer >= 0. This causes problems in conversion from
+ * signed to unsigned integers with -Wconversion. It would be better
+ * to rewrite the functions to return bool and have a pointer where to
+ * store the result as an additional agrument. */
 int canid2ecuid(struct macan_ctx *ctx, uint32_t canid)
 {
 	int i;
@@ -38,7 +43,7 @@ int canid2ecuid(struct macan_ctx *ctx, uint32_t canid)
 
 void print_hexn(const void *data, size_t len)
 {
-	int i;
+	size_t i;
 	const uint8_t *d = data;
 
 	for (i = 0; i < len; i++) {
@@ -51,7 +56,7 @@ void print_hexn(const void *data, size_t len)
 
 int memchk(uint8_t *a, uint8_t *b, size_t len)
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i < len; i++) {
 		if (a[i] != b[i])
@@ -77,7 +82,7 @@ void eval(const char *tname, int b)
 /**
  * memcpy_bw() - memcpy byte-wise */
 void memcpy_bw(void *dst, const void *src, size_t len) {
-        int i;
+        size_t i;
         uint8_t *dst8 = dst;
         const uint8_t *src8 = src;
